requirements.txt
# -*- coding: utf-8 -*-
Django==2.1.5
django-environ==0.4.4
mysqlclient==1.3.13
django_mysql==2.3.0
gunicorn==19.8.1
pymysql==0.9.3
django-model-utils==3.1.2
Fabric3